Test creator.
This commit creates a test creator module that will create the
tests for problems. It takes in the module name and the test
data which is an object that contains two arrays under the
properties input and ouput. The test creator will create a
complie test for the module and then inovke the module with the
inputs and make sure the result matches the outputs. This should
be helpful in creating tests. This will avoid code repition and
speed up testing time.