Queens attack 2 setup.
This commit setups the solution for the queens attack 2 problem.
This commit creates the function, documents everything, solves
the trival cases, and creates a test file for the trival cases.
I think I have somewhat of an idea to count the horizontal options.
I am wondering if I should do it all with numbers or actually build
out the board per say. It might be easier to build out the board in
arrays and then use array opperations to remove squares that aren't
available.