Rook moves.
This commit updates the counting of the number of possible spaces
but including the logic of a rook. This commit implements all the
changes necessary to count for the number of upward, downward,
leftward, and rightward moves taking into account obstacles. I
still need to add in diagonal logic and this problem will be
solved. I am hoping I can do that in 1 loop of the obstacle and
the time complexity would be O(n).