Updating test-creator.
This commit updates the test creator to be able to handle
objects and arrays. The older version of test creator would
fail because it would try to equal the two objects and arrays
which would fail. Deep equal might have been able to solve this,
but checking the keys or values works as well.