Grading students.
This commit implements the solution to the Hackerrank question
of grading students. The details or listed in comments but
essentially grades below 38 are not rounded and are failing.
Grades greater than or equal to 38 can be rounded to the nearest
multiple of 5 if the are less than 3 from that multiple of 5.