Camel case.
This commit implements a solution to the camel case problem. The
probelm was to count the number of words in a camel case string.
I did this by looping through all the characters of the string
and seeing if the ASCII value of the character was an uppercase
letter. If it was I incremented a count. The count started at
1 encase no uppercase letters were found.