Strong password.
This commit implements a solution to the strong password problem.
The module returns the minimum number of characters that need to
be added to a password in order for it to be considered a strong
password. A strong password is one that is atleast 6 characters,
has a lowercase, uppercase, number, and a special character.