Sock merchant.
This commit implements a solution to the sock merchant problem.
Given an array of numbers where each number represents a single
color of a sock, figure out how many pairs of socks can be made.
I simply loop through the colors and have an array where I keep
track of colors that were already seen. If it was already seen
and the current color matches, I remove it from already seen
and increment the count. If it hasn't been already seen, I push
it into the array.