Counting valleys.
This commit implements a solution to the counting valleys
problem. Given a string of all the steps that are taken during
a hike and each step denotes if the step was uphill or downhill,
this function returns the number of valleys in the hike. The
function takes in a string of uppercase 'D's and uppercase 'U's.
'D's denoting downhill and 'U's denoting uphill.