Jumping on the clouds.
This commit implements a solution to the jumping on clouds
problem. An array of zeros and ones are passed into the module.
Ones represent clouds that can't be jumped on and zeros
represent clouds that can be jumped on. The module can take
single jumps or double jumps. The array passed in must always
be solvable. The module returns the minimum number of jumps
that need to be taken.