Array left rotation.
This commit implements a solution to the array left rotation
problem. Given an array preform n left rotations on the array
and return the correct array. A left rotation is such that the
first element becomes the last element and all other elements
are shifted down. So, [1,2,3,4,5] becomes [2,3,4,5,1] are 1
left rotation.