Kangaroo.
This commit implements a solution to the kangaroo problem where
we have two kangaroos and given their intial starting locations
and their velocites do they end up landing on the same spot?
I came up with a soultion that was a brute force one but it did
end up working. I couldn't figure out how to generalize the
solution to consant time or some formula that could be used.