Apples and oranges.
This commit implements a solution to the apples and oranges
problem. Essentailly give 4 locations and distances the apples
and oranges value calculate how many apples and oranges fall
on the tree. We are given the apple trees location, where the
house starts, where the house ends, and the orange trees
location.