Making anagrams.
This commit implements a O(n) solution to the making anagrams
problem. The problem is that given two strings calculate
the minimum number of characters that can be removed from
both strings in order to make the strings anagrams of each
other. My solution looks at the frequencies of the characters
in both strings and then counts how many of each character
needs to be removed.