Special string again.
This commit implements a solution to the special string problem
from the Hackerrank interview prep problems. This soultion
wont pass on hackerrank because it takes too long to execute
on their larger tests sets so it needs to be improved on. This
was my first way to solve it. A special string is a string that
all characters are the same or all characters are the same except
for the middle charater. Give a string we have to calculat how
many special sub strings there are in a string.