Stock Maximize.
This commit implements a solution to the stock maximize
Hackerrank problem. Given an array of future stock prices,
calculate the maximum profit that can be made. Once I finally
broke down the problem, all I needed to find were days were
there is a future day where the stock price is higher than it
is now. Buy stock on those days and sell the stock on the future
day.