Implements solution to Equalize Array Problem.
Determine the minimum number of deletions required in order
to make an array of nubmers contain all of the same element.
I solved this by finiding the number that appeared the most
and subtracting how many times it appeared from the total
length of the array. This results in the minimum number of
deletions that are neccessary to make the array be of the
same element.